Made all text uniform and it all gets larger or smaller when the Font Size option is changed.
Added a placeholder Caregiver Mode check that will limit the amount of options that are displayed if not in Caregiver mode.

Added a temporary "Exit" button to disable Caregiver Mode.

Enabling and disabling Caregiver mode should reload the screen to update the correct amount of settings.

Refactored some repetitive code.